Move getTimeSeparator function to separate file
